+++ src/pages/categories/components/CategoryList.js
@@ -1,43 +1,12 @@
 import React from 'react';
-import { Popconfirm, Button, Table, Space, Form, Select, Input } from 'antd';
+import { Popconfirm, Button, Table, Space } from 'antd';
 
-import { useDispatch, useSelector } from 'react-redux';
-import { getCategories, deleteCategory } from '../../../actions/categories';
+import { useDispatch } from 'react-redux';
+import { deleteCategory } from '../../../actions/categories';
 import { Link } from 'react-router-dom';
-import deepEqual from 'deep-equal';
 
-function CategoryList({ actions }) {
+function CategoryList({ actions, data, filters, setFilters, fetchCategories }) {
   const dispatch = useDispatch();
-  const [filters, setFilters] = React.useState({
-    page: 1,
-    limit: 20,
-  });
-  const [form] = Form.useForm();
-  const { Option } = Select;
-
-  const { categories, total, loading } = useSelector((state) => {
-    const node = state.categories.req.find((item) => {
-      return deepEqual(item.query, filters);
-    });
-
-    if (node)
-      return {
-        categories: node.data.map((element) => state.categories.details[element]),
-        total: node.total,
-        loading: state.categories.loading,
-      };
-    return { categories: [], total: 0, loading: state.categories.loading };
-  });
-
-  React.useEffect(() => {
-    fetchCategories();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [filters]);
-
-  const fetchCategories = () => {
-    dispatch(getCategories(filters));
-  };
-
   const columns = [
     { title: 'Name', dataIndex: 'name', key: 'name' },
     { title: 'Slug', dataIndex: 'slug', key: 'slug' },
@@ -77,37 +46,14 @@ function CategoryList({ actions }) {
 
   return (
     <Space direction={'vertical'}>
-      <Form
-        initialValues={filters}
-        form={form}
-        name="filters"
-        layout="inline"
-        onFinish={(values) => setFilters({ ...filters, ...values })}
-        style={{ maxWidth: '100%' }}
-      >
-        <Form.Item name="q" label="Search" style={{ width: '25%' }}>
-          <Input placeholder="search categories" />
-        </Form.Item>
-        <Form.Item name="sort" label="Sort" style={{ width: '15%' }}>
-          <Select>
-            <Option value="desc">Latest</Option>
-            <Option value="asc">Old</Option>
-          </Select>
-        </Form.Item>
-        <Form.Item>
-          <Button type="primary" htmlType="submit">
-            Submit
-          </Button>
-        </Form.Item>
-      </Form>
       <Table
         bordered
         columns={columns}
-        dataSource={categories}
-        loading={loading}
+        dataSource={data.categories}
+        loading={data.loading}
         rowKey={'id'}
         pagination={{
-          total: total,
+          total: data.total,
           current: filters.page,
           pageSize: filters.limit,
           onChange: (pageNumber, pageSize) =>
