+++ src/pages/tags/components/TagList.js
@@ -1,42 +1,11 @@
 import React from 'react';
-import { Popconfirm, Button, Table, Space, Form, Select, Input } from 'antd';
-import { useDispatch, useSelector } from 'react-redux';
-import { getTags, deleteTag } from '../../../actions/tags';
+import { Popconfirm, Button, Table, Space } from 'antd';
+import { useDispatch } from 'react-redux';
+import { deleteTag } from '../../../actions/tags';
 import { Link } from 'react-router-dom';
-import deepEqual from 'deep-equal';
 
-function TagList({ actions }) {
+function TagList({ actions, filters, setFilters, fetchTags, data }) {
   const dispatch = useDispatch();
-  const [filters, setFilters] = React.useState({
-    page: 1,
-    limit: 20,
-  });
-  const [form] = Form.useForm();
-  const { Option } = Select;
-
-  const { tags, total, loading } = useSelector((state) => {
-    const node = state.tags.req.find((item) => {
-      return deepEqual(item.query, filters);
-    });
-
-    if (node)
-      return {
-        tags: node.data.map((element) => state.tags.details[element]),
-        total: node.total,
-        loading: state.tags.loading,
-      };
-    return { tags: [], total: 0, loading: state.tags.loading };
-  });
-
-  React.useEffect(() => {
-    fetchTags();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [filters]);
-
-  const fetchTags = () => {
-    dispatch(getTags(filters));
-  };
-
   const columns = [
     { title: 'Name', dataIndex: 'name', key: 'name', width: '15%' },
     { title: 'Slug', dataIndex: 'slug', key: 'slug', width: '15%' },
@@ -76,42 +45,14 @@ function TagList({ actions }) {
 
   return (
     <Space direction={'vertical'}>
-      <Form
-        initialValues={filters}
-        form={form}
-        name="filters"
-        layout="inline"
-        onFinish={(values) =>
-          setFilters({
-            ...filters,
-            ...values,
-          })
-        }
-        style={{ maxWidth: '100%' }}
-      >
-        <Form.Item name="q" label="Search" style={{ width: '25%' }}>
-          <Input placeholder="search tags" />
-        </Form.Item>
-        <Form.Item name="sort" label="Sort" style={{ width: '15%' }}>
-          <Select>
-            <Option value="desc">Latest</Option>
-            <Option value="asc">Old</Option>
-          </Select>
-        </Form.Item>
-        <Form.Item>
-          <Button type="primary" htmlType="submit">
-            Submit
-          </Button>
-        </Form.Item>
-      </Form>
       <Table
         bordered
         columns={columns}
-        dataSource={tags}
-        loading={loading}
+        dataSource={data.tags}
+        loading={data.loading}
         rowKey={'id'}
         pagination={{
-          total: total,
+          total: data.total,
           current: filters.page,
           pageSize: filters.limit,
           onChange: (pageNumber, pageSize) =>
